nba-analytics-finale/requirements.txt
# Streaming
kafka-python==2.0.2

# Big Data
pyspark==3.5.0
fastavro==1.9.3

# Machine Learning
scikit-learn==1.3.2
joblib==1.3.2

# RAG
chromadb==0.4.18
sentence-transformers==2.2.2

# LLM (opzionale)
groq==0.4.2

# Dashboard
streamlit==1.29.0
plotly==5.18.0

# Data
pandas==2.1.3
numpy==1.26.2
pyarrow==14.0.1

# Utils
python-dotenv==1.0.0
